requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
coverage==5.0
Django==3.0.1
django-autocomplete-light==3.5.1
django-tempus-dominus==5.1.2.9
django-widget-tweaks==1.4.5
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytz==2019.3
reportlab==3.5.34
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
